WeakHashingAnalyzer audit tests moved to a separate file
