+++ SecurityCodeScan.Test/SecurityCodeScan.Test.csproj
@@ -308,6 +308,7 @@
     <Compile Include="Tests\AspxAnalyzerTest.cs" />
     <Compile Include="Tests\Audit\AuditTest.cs" />
     <Compile Include="Tests\Audit\PathTraversalAnalyzerAuditTest.cs" />
+    <Compile Include="Tests\Audit\WeakHashingAnalyzerAuditTest.cs" />
     <Compile Include="Tests\Config\ConfigurationTest.cs" />
     <Compile Include="Tests\Config\ConfigTests.cs" />
     <Compile Include="Tests\Config\CsrfTokenAnalyzerConfigTests.cs" />
